fix: move test util to mocks
